Adds support for running tests against Postgresql
Adds the necessary plumbing to run the test suite against Postgresql instead of
the current default Derby.  Defining the 'scala-migrations.db.vendor' property
as 'postgresql' switches the tests to use Postgresql.

Additionally fixes a bug with granting permissions when using Postgresql.
Granting usage of the schema is required before the user can take advantage of
any permissions that are granted.